% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_directions.R
\name{local_directions}
\alias{local_directions}
\title{Pass local directions to movemee functions}
\usage{
local_directions(...)
}
\arguments{
\item{...}{commas separated directions. Supports splicing (!!!) vectors in}
}
\value{
A character string double quoted for passage to the terminal
}
\description{
This is a convenience function
all it does is wrap the call in shQuotes
So that it can be passed to the terminal
"shell" safely even with spaces in it
}
