% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_file.R
\name{move_file}
\alias{move_file}
\title{Move file(s) to or from Sharepoint}
\usage{
move_file(
  file,
  destination,
  rclone_command = "rclone copy",
  rclone_options = "--progress --ignore-size",
  show_command_line_command = FALSE
)
}
\arguments{
\item{file}{filepath to file you want moved. Use here() for local files, sharepoint_directions() for sharepoint files}

\item{destination}{filepath to folder you want file moved to. Use here() for local destinations, sharepoint_directions() for sharepoint destinations.}

\item{rclone_options}{options to send to rclone. advanced use only}

\item{show_command_line_command}{optionally see the command run in the terminal}
}
\value{
rclone program output on file transfer status
}
\description{
Move file(s) to or from Sharepoint
}
\examples{
\dontrun{


output_location <- sharepoint_directions('Epidemiology','Surveillance','Random Analysis','Output Folder')

move_file(
here('output','final_graph.png'),
output_location
)

}
}
