% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_working_directory.R
\name{optional_working_directory}
\alias{optional_working_directory}
\title{Decide whether to append the working directory to a filepath}
\usage{
optional_working_directory(x)
}
\arguments{
\item{x}{a string, a filepath}
}
\value{
a string, either the same filepath, or wd/filepath
}
\description{
This function looks for the presence of the working directory
or a colon (indicating a sharepoint location)
If it finds none, it appends the working directory
}
\details{
This is a safeguard - best practice is to use the 'here' package
}
