% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_directions.R
\name{remote_directions}
\alias{remote_directions}
\title{Give directions to any remote file or location
renaming the sharepoint directions for the cases where we're not using a sharepoint link
As of ver .9005 supports splicing (!!!) vectors in}
\usage{
remote_directions(remote_link, ...)
}
\arguments{
\item{remote_link}{The name of the rclone remote link}

\item{...}{subsequent subfolders and filename strings, comma separated}
}
\value{
an appropriately formatted filepath for specifying remote locations in rclone
}
\description{
Give directions to any remote file or location
renaming the sharepoint directions for the cases where we're not using a sharepoint link
As of ver .9005 supports splicing (!!!) vectors in
}
\examples{
\dontrun{
# made up location
remote_directions('Epidemiology','Surveillance','Giardia','Quarterly Report','Input','2023-Q4_report_data.csv')
}

}
