% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharepoint_directions.R
\name{sharepoint_directions}
\alias{sharepoint_directions}
\title{Give directions to a sharepoint file or location
As of ver .9005 supports splicing (!!!) vectors in}
\usage{
sharepoint_directions(remote_link, ...)
}
\arguments{
\item{remote_link}{The name of the Sharepoint site e.g. Epidemiology}

\item{...}{subsequent subfolders and filename strings, comma separated}
}
\value{
an appropriately formatted filepath for specifying Sharepoint locations in rclone
}
\description{
Give directions to a sharepoint file or location
As of ver .9005 supports splicing (!!!) vectors in
}
\examples{
\dontrun{
# made up location
sharepoint_directions('Epidemiology','Surveillance','Giardia','Quarterly Report','Input','2023-Q4_report_data.csv')
}

}
