% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_me_files.R
\name{show_me_files}
\alias{show_me_files}
\title{Show the contents of a sharepoint directory}
\usage{
show_me_files(location, rclone_options = NULL)
}
\arguments{
\item{location}{the sharepoint location you want to know about}

\item{rclone_options}{options to pass to rclone}
}
\value{
dataframe of filenames. Not recursive by default.
}
\description{
Show the contents of a sharepoint directory
}
